Split monitor into a module vs binary for Terminal
Expose the types we need.

Cleanup toml files
