+++ Cargo.toml
@@ -13,6 +13,7 @@ members = [
     "./irq",
     "./memory",
     "./monitor",
+    "./monitor_tui",
 ]
 default-members = [
     "./assemble",
@@ -28,6 +29,7 @@ default-members = [
     "./irq",
     "./memory",
     "./monitor",
+    "./monitor_tui",
 ]
 
 [profile.test]
@@ -40,14 +42,23 @@ debug = true
 debug = true
 
 [workspace.dependencies]
+assemble = { path = "./assemble", version = "~0.1" }
+c64basic = { path = "./c64basic", version = "~0.1" }
+chip = { path = "./chip", version = "~0.1" }
 clap = { version = "~4.0", features = ["derive", "wrap_help"] }
 clap-num = "1.0.2"
 color-eyre = { version = "0.6.2", default-features = false }
+cpu_proc_macros = { path = "./cpu_proc_macros", version = "~0.1" }
+io = { path = "./io", version = "~0.1" }
+irq = { path = "./irq", version = "~0.1" }
+memory = { path = "./memory", version = "~0.1" }
+monitor = { path="./monitor", version = "~0.1" }
 # Force bind proc-macro2 as the default most people are on (1.0.58) doesn't
 # work on nightly after 2023-06-27
 proc-macro2 = { version = "=1.0.66", features=["default", "proc-macro"] }
 rand = "0.8.5"
 regex = "1"
+rusty6502 = { path = "./", version = "~0.1" }
 strum_macros = "0.24.1"
 strum = "0.24.1"
 thiserror = "1.0.44"
