+++ assemble/Cargo.toml
@@ -6,7 +6,7 @@ authors.workspace = true
 rust-version.workspace = true
 
 [dependencies]
-rusty6502  = { path = "../", version = "~0.1" }
+rusty6502  = { workspace = true }
 
 color-eyre = { workspace = true }
 proc-macro2 = { workspace = true }
