+++ assembler/Cargo.toml
@@ -6,9 +6,9 @@ authors.workspace = true
 rust-version.workspace = true
 
 [dependencies]
-assemble = { path = "../assemble", version = "~0.1" }
-rusty6502  = { path = "../", version = "~0.1" }
+rusty6502  = { workspace = true }
 
+assemble = { workspace = true }
 color-eyre = { workspace = true }
 clap = { workspace = true }
 proc-macro2 = { workspace = true }
