+++ cpu/Cargo.toml
@@ -6,11 +6,11 @@ authors.workspace = true
 rust-version.workspace = true
 
 [dependencies]
-chip  = { path = "../chip", version = "~0.1" }
-cpu_proc_macros = { path = "../cpu_proc_macros", version = "~0.1" }
-io  = { path = "../io", version = "~0.1" }
-irq  = { path = "../irq", version = "~0.1" }
-memory  = { path = "../memory", version = "~0.1" }
+chip = { workspace = true }
+cpu_proc_macros = { workspace = true }
+io = { workspace = true }
+irq = { workspace = true }
+memory = { workspace = true }
 
 color-eyre = { workspace = true }
 proc-macro2 = { workspace = true }
